+++ queen/app/api/v1/router.py
@@ -6,6 +6,7 @@ from app.api.v1.endpoints import abdomen, cephalon, nerve, spiracle, thorax
 api_router = APIRouter()
 api_router.include_router(cephalon.router)
 api_router.include_router(cephalon.admin_router)
+api_router.include_router(cephalon.roles_router)
 api_router.include_router(thorax.router)
 api_router.include_router(abdomen.router)
 api_router.include_router(nerve.router)